20-GoogleAppEngine/requirements.txt
Flask==1.1.2
setuptools==40.3.0
google-cloud-ndb == 1.1.2
googleapis_common_protos